+++ src/globals.py
@@ -1,2 +1,2 @@
-build=65
+build=77
 type='stable'